+++ test-e2e/test_calico_networking.py
@@ -11,7 +11,7 @@ from typing import Iterator
 import pytest
 
 from _pytest.fixtures import SubRequest
-from cluster_helpers import wait_for_dcos_oss
+from cluster_helpers import artifact_dir_format, dump_cluster_journals, wait_for_dcos_oss
 from conditional import E2E_SAFE_DEFAULT, escape, only_changed, trailing_path
 from dcos_e2e.backends import Docker
 from dcos_e2e.cluster import Cluster
@@ -40,11 +40,14 @@ def assert_system_unit_state(node: Node, unit_name: str, active: bool = True) ->
 @pytest.fixture(scope="module")
 def calico_ipip_cluster(docker_backend: Docker, artifact_path: Path,
                         request: SubRequest, log_dir: Path) -> Iterator[Cluster]:
+    # Create a relatively large test cluster, since we've seen problems
+    # when many agents attempt to create the Docker network. See
+    # https://jira.d2iq.com/browse/D2IQ-70674
     with Cluster(
             cluster_backend=docker_backend,
-            masters=1,
-            agents=2,
-            public_agents=1,
+            masters=3,
+            agents=8,
+            public_agents=8,
     ) as cluster:
 
         config = {
@@ -70,6 +73,11 @@ def calico_ipip_cluster(docker_backend: Docker, artifact_path: Path,
             log_dir=log_dir,
         )
         yield cluster
+
+        dump_cluster_journals(
+            cluster=cluster,
+            target_dir=log_dir / artifact_dir_format(request.node.name),
+        )
 
 
 @pytest.mark.skipif(